requirements.txt
Flask
openai
python-dotenv
SQLAlchemy
psycopg2-binary
alembic
pytest
psycopg2-binary